requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic-settings==2.12.0
sqlalchemy==2.0.45
alembic==1.18.1
asyncpg==0.31.0
pytest-asyncio==1.3.0
black==25.12.0
isort==7.0.0
flake8==7.3.0
mypy==1.19.1
pre-commit==4.5.1
geoalchemy2==0.18.1
shapely==2.1.2
httpx==0.28.1
